requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.0
gunicorn==25.3.0
idna==3.11
pillow==12.2.0
psycopg2-binary==2.9.11
pycparser==3.0
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
whitenoise==6.12.0
